Add hero search component
The search is done with an objservable that is switched from
the response from hero service searchHeroes. This switched
observable only gets fired if the input holds for 300ms and
only returns the latest values discarding the previous ones,
so if many requests are made only the latest ones are kept.